+++ utils/checkBadWord.ts
